fix big-endian code
